app/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
cerebras-cloud-sdk
chromadb
PyPDF2
python-multipart
requests
sentence-transformers  # For Option 2 (Local Open-Source Model)
openai                 # For Option 1 (OpenAI Embeddings)
mysql-connector-python
pydantic[email]
sqlalchemy